+++ bundles/net.tourbook.tests/src/test/device/suunto/Suunto3_DeviceDataReaderTests.java
@@ -13,7 +13,7 @@
  * this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110, USA
  *******************************************************************************/
-package importdata.suunto3;
+package device.suunto;
 
 import java.util.HashMap;
 
@@ -30,9 +30,9 @@ import org.junit.jupiter.api.Test;
 import utils.Comparison;
 import utils.FilesUtils;
 
-class Suunto3Tests {
+class Suunto3_DeviceDataReaderTests {
 
-   private static final String             IMPORT_FILE_PATH = FilesUtils.rootPath + "importdata/suunto3/files/"; //$NON-NLS-1$
+   private static final String             IMPORT_FILE_PATH = FilesUtils.rootPath + "device/suunto/files/"; //$NON-NLS-1$
 
    private static DeviceData               deviceData;
    private static HashMap<Long, TourData>  newlyImportedTours;
