+++ bundles/net.tourbook.tests/src/test/weather/worldweatheronline/WorldWeatherOnlineTests.java
@@ -13,7 +13,7 @@
  * this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110, USA
  *******************************************************************************/
-package weather.wwo;
+package weather.worldweatheronline;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
